+++ .idea/misc.xml
@@ -23,18 +23,34 @@
         <entry key="../../../../layout/compose-model-1644858162279.xml" value="0.41522917884924126" />
         <entry key="../../../../layout/compose-model-1645523271911.xml" value="0.27294921875" />
         <entry key="../../.gradle/caches/transforms-3/6df8c2a40b2ba565b90d9da7a42b400b/transformed/jetified-firebase-ui-auth-7.2.0/res/drawable/fui_idp_button_background_github.xml" value="0.24791666666666667" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/activity_drawer.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/activity_host.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/activity_login.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/activity_main.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/app_bar_drawer.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/content_drawer.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/fragment_detail.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/fragment_gallery.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/fragment_home.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/fragment_list.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/fragment_map.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/fragment_slideshow.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/fragment_workmates.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/menu_header.xml" value="0.16041666666666668" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/layout/nav_header_drawer.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/menu/activity_main_drawer.xml" value="0.3333333333333333" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/menu/bottom_nav_menu.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/menu/drawer.xml" value="0.29444444444444445" />
+        <entry key="..\:/Work/oc/fadel/Go4Lunch/app/src/main/res/menu/drawer_menu.xml" value="0.29444444444444445" />
         <entry key="app/src/main/res/drawable/ic_baseline_map_24.xml" value="0.2" />
         <entry key="app/src/main/res/drawable/ic_baseline_people_24.xml" value="0.2" />
         <entry key="app/src/main/res/drawable/ic_baseline_view_list_24.xml" value="0.2" />
         <entry key="app/src/main/res/drawable/ic_dashboard_black_24dp.xml" value="0.2" />
         <entry key="app/src/main/res/drawable/ic_home_black_24dp.xml" value="0.2" />
         <entry key="app/src/main/res/font/roboto.xml" value="0.2916666666666667" />
-        <entry key="app/src/main/res/layout/activity_drawer.xml" value="0.3842746896791915" />
-        <entry key="app/src/main/res/layout/activity_host.xml" value="0.3502956761822029" />
+        <entry key="app/src/main/res/layout/activity_host.xml" value="0.1691995947315096" />
         <entry key="app/src/main/res/layout/activity_login.xml" value="0.3499780182570752" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.17933130699088146" />
-        <entry key="app/src/main/res/layout/app_bar_drawer.xml" value="0.22604166666666667" />
-        <entry key="app/src/main/res/layout/content_drawer.xml" value="0.2375" />
         <entry key="app/src/main/res/layout/fragment_dashboard.xml" value="0.15942028985507245" />
         <entry key="app/src/main/res/layout/fragment_detail.xml" value="0.3619791666666667" />
         <entry key="app/src/main/res/layout/fragment_home.xml" value="0.182518115942029" />
@@ -42,14 +58,11 @@
         <entry key="app/src/main/res/layout/fragment_map.xml" value="0.1691995947315096" />
         <entry key="app/src/main/res/layout/fragment_maps.xml" value="0.1" />
         <entry key="app/src/main/res/layout/fragment_notifications.xml" value="0.15942028985507245" />
-        <entry key="app/src/main/res/layout/menu_header.xml" value="0.317481884057971" />
-        <entry key="app/src/main/res/menu/activity_main_drawer.xml" value="0.396875" />
         <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.1" />
-        <entry key="app/src/main/res/menu/drawer_menu.xml" value="0.3828125" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" project-jdk-name="1.8" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
