--- app/src/main/java/com/fadel/go4lunch/HostActivity.kt
@@ -1,56 +1,0 @@
-package com.fadel.go4lunch
-
-import android.os.Bundle
-import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
-import com.fadel.go4lunch.databinding.ActivityHostBinding
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import com.google.firebase.auth.FirebaseAuth
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class HostActivity : AppCompatActivity() {
-
-    private val vm by viewModels<HostActivityViewModel>()
-    private lateinit var binding: ActivityHostBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityHostBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val navView: BottomNavigationView = binding.navView
-
-        val navController = findNavController(R.id.nav_host_fragment_activity_host)
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
-        val appBarConfiguration = AppBarConfiguration(
-            setOf(
-                R.id.navigation_map, R.id.navigation_list, R.id.navigation_workmates
-            )
-        )
-        setupActionBarWithNavController(navController, appBarConfiguration)
-        navView.setupWithNavController(navController)
-    }
-
-    override fun onResume() {
-        super.onResume()
-        vm.onResume()
-    }
-
-    /* private fun initializeUI() {
-         binding.logoutButton.setOnClickListener {
-             logout()
-         }
-     }*/
-
-    /*private fun logout() {
-        FirebaseAuth.getInstance().signOut();
-    }*/
-
-}
