+++ app/src/main/java/com/fadel/go4lunch/ui/list/ListFragment.kt
@@ -89,9 +89,7 @@ class ListFragment : Fragment() {
             viewLifecycleOwner,
             Observer {
                 when (it) {
-                    is ListViewModel.NavigationOrder.Detail -> findNavController().navigate(
-                        R.id.toDetail
-                    )
+                    is ListViewModel.NavigationOrder.Detail -> TODO()
                 }
             }
         )
