+++ app/src/main/java/com/fadel/go4lunch/ui/main/MainViewModel.kt
@@ -1,15 +1,16 @@
-package com.fadel.go4lunch
+package com.fadel.go4lunch.ui.main
 
 import android.Manifest
 import android.app.Application
 import android.content.pm.PackageManager
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.ViewModel
+import com.fadel.go4lunch.data.PermissionRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class HostActivityViewModel @Inject constructor(
+class MainViewModel @Inject constructor(
     private val context: Application,
     private val permissionRepository: PermissionRepository
 ) : ViewModel() {
