--- app/src/main/java/com/fadel/go4lunch/ui/ui/slideshow/SlideshowViewModel.kt
@@ -1,18 +1,0 @@
-package com.fadel.go4lunch.ui.ui.slideshow
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import dagger.hilt.android.lifecycle.HiltViewModel
-import javax.inject.Inject
-
-@HiltViewModel
-class SlideshowViewModel @Inject constructor(
-
-): ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
-    }
-    val text: LiveData<String> = _text
-}
