+++ app/src/main/res/layout/fragment_list.xml
@@ -7,4 +7,11 @@
     tools:context=".ui.list.ListFragment">
 
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        android:text="list"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
