+++ app/src/main/res/layout/fragment_map.xml
@@ -6,4 +6,11 @@
     android:layout_height="match_parent"
     tools:context=".ui.map.MapFragment">
 
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        android:text="map"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
