--- app/src/main/res/navigation/mobile_navigation.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/mobile_navigation"
-    app:startDestination="@+id/navigation_map">
-
-    <fragment
-        android:id="@+id/navigation_map"
-        android:name="com.fadel.go4lunch.ui.map.MapFragment"
-        android:label="@string/title_map"
-        tools:layout="@layout/fragment_map" />
-
-    <fragment
-        android:id="@+id/navigation_list"
-        android:name="com.fadel.go4lunch.ui.list.ListFragment"
-        android:label="@string/title_list"
-        tools:layout="@layout/fragment_list" >
-        <action
-            android:id="@+id/toDetail"
-            app:destination="@id/navigation_detail" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_detail"
-        android:name="com.fadel.go4lunch.ui.list.ListFragment"
-        android:label="@string/title_list"
-        tools:layout="@layout/fragment_detail" >
-        <argument
-            android:name="id"
-            app:argType="string" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_workmates"
-        android:name="com.fadel.go4lunch.ui.workmates.WorkmatesFragment"
-        android:label="@string/title_workmates"
-        tools:layout="@layout/fragment_workmates" />
-</navigation>
